feat(location): raise exception in irregular cases

+++ MachineAPI/src/Application/Services/LocationService.cs
@@ -25,24 +25,51 @@ namespace MachineAPI.Application.Services
 
         public async Task<LocationDTO?> GetById(int id)
         {
-            var location = await _locationRepository.GetById(id);
+            if (id <= 0)
+            {
+                throw new ArgumentException("ID inválido.");
+            }
+
+            Location? location = await _locationRepository.GetById(id);
+
+            if (location == null)
+            {
+                throw new KeyNotFoundException($"Localização com ID {id} não encontrada.");
+            }
+
             return location != null ? _mapper.Map<LocationDTO>(location) : null;
         }
 
         public async Task Add(LocationDTO locationDTO)
         {
+            if (locationDTO == null)
+            {
+                throw new ArgumentNullException(nameof(locationDTO), "Nenhum dado foi recebido.");
+            }
+
             Location location = _mapper.Map<Location>(locationDTO);
             await _locationRepository.Add(location);
         }
 
         public async Task Update(int id, LocationDTO locationDTO)
         {
+            if (id <= 0)
+            {
+                throw new ArgumentException("ID inválido.");
+            }
+
+            if (locationDTO == null)
+            {
+                throw new ArgumentNullException(nameof(locationDTO), "Nenhum dado foi recebido.");
+            }
+
             Location? location = await _locationRepository.GetById(id);
 
             if (location == null)
             {
-                throw new KeyNotFoundException($"Category with ID {id} not found.");
+                throw new KeyNotFoundException($"Localização com ID {id} não encontrada.");
             }
+
             _mapper.Map(locationDTO, location);
 
             await _locationRepository.Update(location);
@@ -50,6 +77,18 @@ namespace MachineAPI.Application.Services
 
         public async Task Delete(int id)
         {
+            if (id <= 0)
+            {
+                throw new ArgumentException("ID inválido.");
+            }
+
+            Location? location = await _locationRepository.GetById(id);
+
+            if (location == null)
+            {
+                throw new KeyNotFoundException($"Localização com ID {id} não encontrada.");
+            }
+
             await _locationRepository.Delete(id);
         }
     }
